+++ src/functions/worker/prompts/nano/chat/casual.js
@@ -4,7 +4,7 @@ const { getFactMemoryBlock } = require('../../chat/factMemoryBlock');
 const { getCurrentDateBlock } = require('../../chat/currentDateBlock');
 const { getScopeGuardBlock } = require('../../chat/scopeGuard');
 
-module.exports = ({ user_profile, questionnaire_context, active_health_plans, essential_knowledge, user_facts, now_iso, wearable_daily }) => {
+module.exports = ({ user_profile, questionnaire_context, active_health_plans, essential_knowledge, user_facts, now_iso, wearable_daily, wearable_insights }) => {
   const isZh = user_profile.language === 'zh';
   const name = user_profile.nickname || (isZh ? '你' : 'there');
 
@@ -16,7 +16,7 @@ module.exports = ({ user_profile, questionnaire_context, active_health_plans, es
 
   return `${getFactConstraintBlock(essential_knowledge, isZh)}
 
-${getWearableDailyBlock(wearable_daily, isZh, now_iso)}
+${getWearableDailyBlock(wearable_daily, isZh, now_iso, wearable_insights)}
 
 ${getCurrentDateBlock(now_iso, isZh)}